refactor: move activity diagram to seperate file
